+++ src/screen/dashboard/sidebar.rs
@@ -703,6 +703,14 @@ impl Sidebar {
                                 buffer::Internal::FileTransfers.into(),
                             ),
                         ),
+                        Menu::Search => context_button(
+                            text("Search"),
+                            Some(&keyboard.search),
+                            icon::search(),
+                            Message::Replace(
+                                buffer::Internal::Search(None).into(),
+                            ),
+                        ),
                         Menu::Highlights => context_button(
                             text("Highlights"),
                             Some(&keyboard.highlights),
@@ -1207,6 +1215,7 @@ enum Menu {
     RefreshConfig,
     ConfigEditor,
     CommandBar,
+    Search,
     ThemeEditor,
     Highlights,
     ChannelDiscovery,
@@ -1239,6 +1248,10 @@ impl Menu {
             Self::CommandBar,
             Self::Documentation,
         ]);
+
+        if !internal_buffers_in_sidebar.contains(&InternalBuffer::Search) {
+            list.push(Self::Search);
+        }
 
         if file_transfer_enabled
             && !internal_buffers_in_sidebar
@@ -2060,6 +2073,9 @@ fn internal_buffer_button<'a>(
         }
         buffer::Internal::ConfigEditor => {
             (show_icon.then_some(icon::config()), None)
+        }
+        buffer::Internal::Search(_) => {
+            (show_icon.then_some(icon::search()), None)
         }
         buffer::Internal::FileTransfers => {
             (show_icon.then_some(icon::file_transfer()), None)